Change prop EngineVolume int to double and add validation for it

+++ CarPrice/Models/Car.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace CarPrice.Models
 {
     internal sealed record Car(string Company, string Model, int Mileage, 
-                               int EnginePower, int EngineVolume, string Year, 
+                               int EnginePower, double EngineVolume, string Year, 
                                bool Transmission, int Price = 0)
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
